Add RunAddsTests parameter and update testing pipeline
Introduced a new `RunAddsTests` parameter in `app-deploy.yml` with options for conditional testing. Added a job template for `TriggerADDSAutoTestsPipeline` that activates based on the `RunAddsTests` value. Updated `azure-pipelines.yml` to include a new repository for the testing pipeline and set `RunAddsTests` to 'vNextIAT' in multiple stages, enhancing the deployment process with more flexible testing capabilities.